verifier tests should be in filter tests
